+++ app/src/main/kotlin/com/farminos/print/Service.kt
@@ -95,10 +95,10 @@ data class PrinterWithSettingsAndInfo(val printer: Printer, val settings: Printe
 
 fun buildPrinterInfo(id: PrinterId, name: String, settings: PrinterSettings): PrinterInfo {
     val dpi = settings.dpi
-    val width = settings.width.toDouble()
-    val height = settings.height.toDouble()
+    val width = settings.width
+    val height = settings.height
     val marginCm = settings.marginCm
-    val marginMils = cmToMils(marginCm.toDouble())
+    val marginMils = cmToMils(marginCm)
     val df = DecimalFormat("#.#")
     val mediaSizeLabel = "${df.format(width)}x${df.format(height)}cm"
     return PrinterInfo.Builder(id, name, PrinterInfo.STATUS_IDLE)
